Fixed some bugs, added comments, and added image cache
