+++ Weather/Models/Forecast.swift
@@ -24,6 +24,7 @@ struct LocationInformation: Codable {
 }
 
 // Struct for the Forecast for one location
+// API examples didn't specify what was optional or not, and the examples wasn't consistent, so made assumptions about what is optional and not
 struct Forecast: Codable {
     let coord: Coordinates
     let weather: [Weather]
@@ -40,6 +41,9 @@ struct Forecast: Codable {
     let id: Int
     let name: String
     let cod: Int?
+    // These properties help load data when starting the app from launch
+    // save image as Data because Codable doesn't like when a property is UIImage
     var image: Data?
+    // Also added this one to hold whether to display Current Location if no city was found by coords or to display city name with state since API doesn't give state with the name property
     var cityStateOrCurrentLocation: String?
 }
